+++ python-libraries/data-platform-catalogue/tests/client/datahub/test_graphql_helpers.py
@@ -1,4 +1,8 @@
-from data_platform_catalogue.client.datahub.graphql_helpers import parse_columns
+from data_platform_catalogue.client.datahub.graphql_helpers import (
+    parse_columns,
+    parse_relations,
+)
+from data_platform_catalogue.entities import RelatedEntity, RelationshipType
 
 
 def test_parse_columns_with_primary_key_and_foreign_key():
@@ -118,3 +122,31 @@ def test_parse_columns_with_no_schema():
 
     assert parse_columns(entity) == []
     assert parse_columns(entity) == []
+
+
+def test_parse_relations():
+    relations = {
+        "relationships": {
+            "total": 1,
+            "relationships": [
+                {
+                    "entity": {
+                        "urn": "urn:li:dataProduct:test",
+                        "properties": {"name": "test"},
+                    }
+                }
+            ],
+        }
+    }
+    result = parse_relations(RelationshipType.PARENT, relations["relationships"])
+    assert result == {
+        RelationshipType.PARENT: [
+            RelatedEntity(id="urn:li:dataProduct:test", name="test")
+        ]
+    }
+
+
+def test_parse_relations_blank():
+    relations = {"relationships": {"total": 0, "relationships": []}}
+    result = parse_relations(RelationshipType.PARENT, relations["relationships"])
+    assert result == {RelationshipType.PARENT: []}